requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
packaging==25.0
django==4.2.16
django-cors-headers==4.3.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==6.2.0
requests==2.32.3
rq==2.4.1
six==1.17.0
sqlalchemy==2.0.31
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
python-dotenv==1.0.0
pandas==2.2.2
lxml==5.2.2
Flask==3.0.3
Flask-Cors==4.0.0
